Time: 2 ms (80.00%), Space: 6.8 MB (51.75%) - LeetHub

+++ 2311-longest-binary-subsequence-less-than-or-equal-to-k/2311-longest-binary-subsequence-less-than-or-equal-to-k.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int longestSubsequence(string s, int k) {
         int n=s.size();
-        long long val=0,pow=1;
+        int val=0,pow=1;
         int cnt=0;
         for(int i=n-1;i>=0 && val+pow<=k;i--){
             if(s[i]=='1'){
